requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
outscraper==3.0.8
python-dotenv==1.0.0